requirements.txt
discord.py
flask
aiohttp
aiohttp